pool: move payment scheme helpers to paymentmgr.
This makes payment scheme helpers paymentmgr methods, helpers that use share id keys have been updated to reference the new share id structure.
Associated tests have been updated and moved
as well.

Difficulty related helpers have also been moved
to their appropriate location (in difficulty.go).
