+++ pool/difficulty.go
@@ -2,6 +2,7 @@ package pool
 
 import (
 	"fmt"
+	"math"
 	"math/big"
 	"sync"
 	"time"
@@ -31,6 +32,58 @@ var (
 		WhatsminerD1:  new(big.Int).SetInt64(48e12),
 	}
 )
+
+// DifficultyToTarget converts the provided difficulty to a target based on the
+// active network.
+func DifficultyToTarget(net *chaincfg.Params, difficulty *big.Rat) (*big.Rat, error) {
+	powLimit := new(big.Rat).SetInt(net.PowLimit)
+
+	// The corresponding target is calculated as:
+	//
+	//    target = pow_limit / difficulty
+	//
+	// The result is clamped to the pow limit if it exceeds it.
+	target := new(big.Rat).Quo(powLimit, difficulty)
+	if target.Cmp(powLimit) > 0 {
+		target = powLimit
+	}
+	return target, nil
+}
+
+// calculatePoolDifficulty determines the difficulty at which the provided
+// hashrate can generate a pool share by the provided target time.
+func calculatePoolDifficulty(net *chaincfg.Params, hashRate *big.Int, targetTimeSecs *big.Int) *big.Rat {
+	hashesPerTargetTime := new(big.Int).Mul(hashRate, targetTimeSecs)
+	powLimit := net.PowLimit
+	powLimitFloat, _ := new(big.Float).SetInt(powLimit).Float64()
+
+	// The number of possible iterations is calculated as:
+	//
+	//    iterations := 2^(256 - floor(log2(pow_limit)))
+	iterations := math.Pow(2, 256-math.Floor(math.Log2(powLimitFloat)))
+
+	// The difficulty at which the provided hashrate can mine a block is
+	// calculated as:
+	//
+	//    difficulty = (hashes_per_sec * target_in_seconds) / iterations
+	difficulty := new(big.Rat).Quo(new(big.Rat).SetInt(hashesPerTargetTime),
+		new(big.Rat).SetFloat64(iterations))
+
+	// Clamp the difficulty to 1 if needed.
+	if difficulty.Cmp(ZeroRat) == 0 || difficulty.Cmp(ZeroRat) < 0 {
+		difficulty = new(big.Rat).SetInt64(1)
+	}
+	return difficulty
+}
+
+// calculatePoolTarget determines the target difficulty at which the provided
+// hashrate can generate a pool share by the provided target time.
+func calculatePoolTarget(net *chaincfg.Params, hashRate *big.Int, targetTimeSecs *big.Int) (*big.Rat, *big.Rat, error) {
+	difficulty := calculatePoolDifficulty(net, hashRate, targetTimeSecs)
+	target, err := DifficultyToTarget(net, difficulty)
+
+	return target, difficulty, err
+}
 
 // DifficultyInfo represents the difficulty related info for a mining client.
 type DifficultyInfo struct {
