+++ pool/paymentmgr_test.go
@@ -51,6 +51,93 @@ func testPaymentMgr(t *testing.T, db *bolt.DB) {
 		t.Fatalf("[NewPaymentMgr] unexpected error: %v", err)
 	}
 
+	// Test sharePercentages.
+	shareSet := map[string]struct {
+		input  []*Share
+		output map[string]*big.Rat
+		err    error
+	}{
+		"equal shares": {
+			input: []*Share{
+				NewShare("a", new(big.Rat).SetInt64(5)),
+				NewShare("b", new(big.Rat).SetInt64(5)),
+				NewShare("c", new(big.Rat).SetInt64(5)),
+				NewShare("d", new(big.Rat).SetInt64(5)),
+				NewShare("e", new(big.Rat).SetInt64(5)),
+			},
+			output: map[string]*big.Rat{
+				"a": new(big.Rat).SetFrac64(5, 25),
+				"b": new(big.Rat).SetFrac64(5, 25),
+				"c": new(big.Rat).SetFrac64(5, 25),
+				"d": new(big.Rat).SetFrac64(5, 25),
+				"e": new(big.Rat).SetFrac64(5, 25),
+			},
+			err: nil,
+		},
+		"inequal shares": {
+			input: []*Share{
+				NewShare("a", new(big.Rat).SetInt64(5)),
+				NewShare("b", new(big.Rat).SetInt64(10)),
+				NewShare("c", new(big.Rat).SetInt64(15)),
+				NewShare("d", new(big.Rat).SetInt64(20.0)),
+				NewShare("e", new(big.Rat).SetInt64(25.0)),
+			},
+			output: map[string]*big.Rat{
+				"a": new(big.Rat).SetFrac64(5, 75),
+				"b": new(big.Rat).SetFrac64(10, 75),
+				"c": new(big.Rat).SetFrac64(15, 75),
+				"d": new(big.Rat).SetFrac64(20, 75),
+				"e": new(big.Rat).SetFrac64(25, 75),
+			},
+			err: nil,
+		},
+		"zero shares": {
+			input: []*Share{
+				NewShare("a", new(big.Rat)),
+				NewShare("b", new(big.Rat)),
+				NewShare("c", new(big.Rat)),
+				NewShare("d", new(big.Rat)),
+				NewShare("e", new(big.Rat)),
+			},
+			output: nil,
+			err:    MakeError(ErrDivideByZero, "division by zero", nil),
+		},
+	}
+
+	for name, test := range shareSet {
+		actual, err := mgr.sharePercentages(test.input)
+		if err != test.err {
+			var errCode ErrorCode
+			var expectedCode ErrorCode
+
+			if err != nil {
+				e, ok := err.(Error)
+				if ok {
+					errCode = e.ErrorCode
+				}
+			}
+
+			if test.err != nil {
+				e, ok := test.err.(Error)
+				if ok {
+					expectedCode = e.ErrorCode
+				}
+			}
+
+			if errCode.String() != expectedCode.String() {
+				t.Fatalf("%s: error generated was %v, expected %v.",
+					name, errCode.String(), expectedCode.String())
+			}
+		}
+
+		for account, dividend := range test.output {
+			if actual[account].Cmp(dividend) != 0 {
+				t.Fatalf("%s: account %v dividend was %v, "+
+					"expected %v.", name, account, actual[account], dividend)
+			}
+		}
+	}
+
 	// Test pruneShares.
 	now := time.Now()
 	zeroSource := &PaymentSource{
@@ -80,13 +167,15 @@ func testPaymentMgr(t *testing.T, db *bolt.DB) {
 	}
 
 	// Ensure share A got pruned with share B remaining.
-	_, err = fetchShare(db, nanoToBigEndianBytes(minimumTime))
+	shareAID := shareID(xID, minimumTime)
+	_, err = fetchShare(db, shareAID)
 	if err == nil {
 		t.Fatal("expected value not found error")
 	}
 
-	_, err = fetchShare(db, nanoToBigEndianBytes(maximumTime))
-	if err == nil {
+	shareBID := shareID(yID, maximumTime)
+	_, err = fetchShare(db, shareBID)
+	if err != nil {
 		t.Fatalf("unexpected error fetching share B: %v", err)
 	}
 
@@ -98,9 +187,166 @@ func testPaymentMgr(t *testing.T, db *bolt.DB) {
 	}
 
 	// Ensure share B got pruned.
-	_, err = fetchShare(db, nanoToBigEndianBytes(maximumTime))
+	_, err = fetchShare(db, shareBID)
 	if err == nil {
 		t.Fatalf("expected value not found error")
+	}
+
+	// Test PPSEligibleShares and PPLNSEligibleShares.
+	now = time.Now()
+	minimumTime = now.Add(-(time.Second * 60)).UnixNano()
+	maximumTime = now.UnixNano()
+	belowMinimumTime := now.Add(-(time.Second * 80)).UnixNano()
+	aboveMaximumTime = now.Add(time.Second * 10).UnixNano()
+	weight = new(big.Rat).SetFloat64(1.0)
+
+	shareCount := 1
+	expectedShareCount := 2
+
+	// Create a share below the PPS time range for account x.
+	err = persistShare(db, xID, weight, belowMinimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share at minimum of the PPS time range for account x.
+	err = persistShare(db, xID, weight, minimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share at maximum of the PPS time range for account y.
+	err = persistShare(db, yID, weight, maximumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share above maximum of the PPS time range for account y.
+	err = persistShare(db, yID, weight, aboveMaximumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	minimumTimeBytes := nanoToBigEndianBytes(minimumTime)
+	maximumTimeBytes := nanoToBigEndianBytes(maximumTime)
+
+	// Fetch eligible shares using the minimum and maximum time range.
+	shares, err := mgr.PPSEligibleShares(minimumTimeBytes, maximumTimeBytes)
+	if err != nil {
+		t.Fatalf("PPSEligibleShares: unexpected error: %v", err)
+	}
+
+	// Ensure the returned share count is as expected.
+	if len(shares) != expectedShareCount {
+		t.Fatalf("PPS error: expected %v eligible PPS shares, got %v",
+			expectedShareCount, len(shares))
+	}
+
+	forAccX := 0
+	forAccY := 0
+	for _, share := range shares {
+		if share.Account == xID {
+			forAccX++
+		}
+
+		if share.Account == yID {
+			forAccY++
+		}
+	}
+
+	// Ensure account x and account y both have shares returned.
+	if forAccX == 0 || forAccY == 0 {
+		t.Fatalf("PPS error: expected shares for account X and Y, "+
+			"got %v (for x), %v (for y).", forAccX, forAccY)
+	}
+
+	// Ensure account x and account y have equal number of shares.
+	if forAccX != forAccY {
+		t.Fatalf("PPS error: expected equal shares for account X and Y, "+
+			"got %v (for x), %v (for y).", forAccX, forAccY)
+	}
+
+	// Ensure account x and account y have shares equal to the share count.
+	if forAccX != shareCount || forAccY != shareCount {
+		t.Fatalf("PPS error: expected share counts of %v for account X and Y, "+
+			"got %v (for x), %v (for y).", shareCount, forAccX, forAccY)
+	}
+
+	// Empty the share bucket.
+	err = emptyBucket(db, shareBkt)
+	if err != nil {
+		t.Fatalf("emptyBucket error: %v", err)
+	}
+
+	// Create a share below the minimum exclusive PPLNS time for account x.
+	err = persistShare(db, xID, weight, belowMinimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share below the minimum exxcuive PPLNS time for account y.
+	err = persistShare(db, yID, weight, belowMinimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share at minimum exclusive PPLNS time for account x.
+	err = persistShare(db, xID, weight, minimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share at minimum exclusive PPLNS time for account y.
+	err = persistShare(db, yID, weight, minimumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share above minimum exclusive PPLNS time for account x.
+	err = persistShare(db, xID, weight, maximumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a share above minimum exclusive PPLNS time for account y.
+	err = persistShare(db, yID, weight, aboveMaximumTime)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	shares, err = mgr.PPLNSEligibleShares(minimumTimeBytes)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Ensure the returned number of shates is as expected.
+	if len(shares) != expectedShareCount {
+		t.Fatalf("PPLNS error: expected %v eligible PPLNS shares, got %v",
+			expectedShareCount, len(shares))
+	}
+
+	// Ensure account x and account y both have shares returned.
+	if forAccX == 0 || forAccY == 0 {
+		t.Fatalf("PPLNS error: expected shares for account X and Y, "+
+			"got %v (for x), %v (for y).", forAccX, forAccY)
+	}
+
+	// Ensure account x and account y have equal number of shares.
+	if forAccX != forAccY {
+		t.Fatalf("PPLNS error: expected equal shares for account X and Y, "+
+			"got %v (for x), %v (for y).", forAccX, forAccY)
+	}
+
+	// Ensure account x and account y have shares equal to the share count.
+	if forAccX != shareCount || forAccY != shareCount {
+		t.Fatalf("PPLNS error: expected share counts of %v for account X and Y, "+
+			"got %v (for x), %v (for y).", shareCount, forAccX, forAccY)
+	}
+
+	// Empty the share bucket.
+	err = emptyBucket(db, shareBkt)
+	if err != nil {
+		t.Fatalf("emptyBucket error: %v", err)
 	}
 
 	// Test pendingPayments, pendingPaymentsAtHeight,
@@ -347,7 +593,7 @@ func testPaymentMgr(t *testing.T, db *bolt.DB) {
 	now = time.Now()
 	sixtyBefore := now.Add(-(time.Second * 60)).UnixNano()
 	thirtyBefore := now.Add(-(time.Second * 30)).UnixNano()
-	shareCount := 10
+	shareCount = 10
 	coinbaseValue := 80
 	height = uint32(20)
